Fix JUnit Category
