+++ subprojects/performance/src/performanceTest/groovy/org/gradle/performance/regression/inception/GradleInceptionPerformanceTest.groovy
@@ -43,6 +43,7 @@ import static org.gradle.test.fixtures.server.http.MavenHttpPluginRepository.PLU
  *   - e.g. change in Gradle that breaks the Gradle build
  */
 @Issue('https://github.com/gradle/gradle-private/issues/1313')
+@Category(SlowPerformanceRegressionTest)
 class GradleInceptionPerformanceTest extends AbstractCrossVersionGradleInternalPerformanceTest {
 
     static List<String> extraGradleBuildArguments() {
@@ -58,7 +59,6 @@ class GradleInceptionPerformanceTest extends AbstractCrossVersionGradleInternalP
         runner.minimumBaseVersion = GradleVersion.version(targetVersion).baseVersion.version
     }
 
-    @Category(SlowPerformanceRegressionTest)
     @Unroll
     def "buildSrc api change in #testProject comparing gradle"() {
         given:
